requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2025.3.1
gast==0.5.4
GDAL==3.4.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.6
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.22.0
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sympy==1.13.1
termcolor==2.4.0
tifffile==2024.4.24
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.18.1
